+++ src/main/js/field/FieldsPanel.js
